backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Capstone Project backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask==3.1.2",
    "pydantic==2.11.8",
    "redis>=6.4.0",
    "geopandas>=1.1.1",
    "pandasql>=0.7.3",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=4.0.0",
    "geoalchemy2>=0.18.0",
    "python-dotenv>=1.0.1",
    "asyncpg>=0.30.0",
    "openai==1.107.2",
    "google-genai==1.36.0",
    "pytest>=8.4.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]

[tool.setuptools.packages.find]
where = ["app"]
